Rename flag variables as `-flag` following elisp coding conventions
From the "Coding Conventions" page of the Emacs manual:

If the purpose of a function is to tell you whether a certain condition
is true or false, give the function a name that ends in ‘p’ (which
stands for “predicate”). If the name is one word, add just ‘p’; if the
name is multiple words, add ‘-p’. Examples are framep and frame-live-p.
We recommend to avoid using this -p suffix in boolean variable names,
unless the variable is bound to a predicate function; instead, use a
-flag suffix or names like is-foo.